Update liballoc lib
This version works but the `LICENSE` seems to be GPLv3 now. I might have
to remove this lib if this is going to be a problem for the rest of the
project. I won't release binaries but I'd like to keep the rest of the
source code more "open" (a.k.a. under the MIT license).