strdup() argv before passing them to k_execv() in kmain()
This fixes a nasty bug: 'argv[0]' (aka 'path' or 'pathname' in
'k_open()') was overwritten by 'vfs_namei_mount()' via 'npath'. The
problem was that 'npath' was pointing to 'path', even after a call to
'strdup()'. This is likely because the heap allocator thought it was
okay to reuse this address.