kernel: mmu: fix frame_deallocate() by looking up the request
This fixes an issue I found with these STR:

1. boot in kernel mode
2. run `cat /proc/meminfo` and write down the number of frames used
3. run `selftest`
4. run `cat /proc/meminfo` and write down the number of frames used
5. run `selftest`
6. run `cat /proc/meminfo` and write down the number of frames used

The number of frames used in (3) and (4) should be very close to each other,
and the number of frames used in (4) and (6) should be identical. Otherwise it
means we are leaking frames, and potentially marking frames in-use as "free" to
re-use again.

This is because the `allocated_frames` is set with `request` values, not
`frame_number`. We use `request` for good reasons (because there are holes in
memory IIRC).

This patch fixes the problem by introducing `frame_number_to_request()`.