build: refactor how docs are generated
It is still Doxygen but it is now themed with
https://github.com/jothepro/doxygen-awesome-css, which is a huuuuge
improvement over the built-in Doxygen theme.

I also found a way to better link the different README.md files by using
HTML comments with special Doxygen directives. It isn't super clean but
it isn't too bad either, and that works.